+++ packages/commitlint-config-cozy/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.17"></a>
+## [0.3.17](https://github.com/cozy/cozy-libs/compare/commitlint-config-cozy@0.3.16...commitlint-config-cozy@0.3.17) (2019-02-11)
+
+
+
+
+**Note:** Version bump only for package commitlint-config-cozy
+
 <a name="0.3.16"></a>
 ## [0.3.16](https://github.com/cozy/cozy-libs/compare/commitlint-config-cozy@0.3.15...commitlint-config-cozy@0.3.16) (2019-01-16)